Refactor FoodController to accept double for calories, enhance MealPlan model with UserId and FoodId, and implement validation in MyContext. Add unit tests for food operations and update project dependencies.

+++ AppHealthyFood/Controllers/FoodController.cs
@@ -23,7 +23,7 @@ namespace AppHealthyFood.Controllers
             return db.Foods.ToList();
         }
 
-        public void UpdateFood(int id, int calories)
+        public void UpdateFood(int id, double calories)
         {
             var food = db.Foods.Find(id);
             if (food != null)
